+++ CHANGELOG.md
@@ -145,3 +145,4 @@
 - 2023-10-24: Pagination deals, session, gp
 - 2023-10-25: Serde deals, session
 - 2023-10-26: Features deals
+- 2023-10-27: Metrics deals, qualifier, gp